move storejson to pca folder
